Updated the name inputbox
Used a hacky solution to have the whole screen clickable and in-focus for a more user-friendly experience. While the old input box used built-in functions and dynamically scaled with the frame, it only allowed input IF the user's mouse was over the input box. While intended for that type of on-screen point & click interface, it was not what I was looking for.